Passing props into colors can be fixed.
Passing each individual color could not previously be fixed because tailwind needs to detect the full class name (such as the full declaration of the text color), not just the color. While this means that props can be passed in, it will not be efficient during the creation of the api for when instead of colors being passed in individually, such as 'blue', they will have to be passed in as 'text-blue-500', for example. The documentation for the fixes to these problems is here (https://tailwindcss.com/docs/content-configuration#class-detection-in-depth), which will assisst in solving the issue.