Changing showLogin to loginType
Current variable for whether or not the login shows / whether the user logs in or signs up depends on a true/false variable. This is now being changed to a variable with 3 possible values: signUp, login, and resetPassword to allow the user to reset their password, log in, or sign Up.