Reset Password page has been implemented.
By suggestion of multiple team members, the reset-password page functionality has been implemented. There is, however, a styling issue to be fixed and an error that is lacking on the login despite the password input fields matching on both pages.

There is a high chance this issue will be overdue. This is due to an error made on my part in attempting to use the login page as the sign Up and "reset" password interface at the same time.